Bug #651187 update the AE to the lasted version
[root cause  ] ae lib need to update
[changes     ] libcamera
[side effects]none
[self test   ]<board, Test cases summary>:<9850ka_1h10, pre and cap is ok>
[reviewers   ] zhidong.li
[change_type ] feature_bugfix
[tag_product ] common

+++ isp2.1/ae/sprd_ae/inc/ae_porting.h
@@ -64,13 +64,13 @@ typedef signed char int8;
 typedef unsigned char boolean;
 	typedef unsigned char uint8_t;
 	typedef unsigned short uint16_t;
-	typedef unsigned long int uint32_t;
+typedef unsigned int uint32_t;
 	typedef unsigned __int64 uint64_t;
 	typedef unsigned int uint_t;
 	
 typedef signed char int8_t;
 	typedef signed short int16_t;
-	typedef signed long int int32_t;
+typedef signed int int32_t;
 	typedef signed __int64 int64_t;
 	
 typedef unsigned __int64 __u64;
